Move StreamField frontend code out of components folder
