+++ client/src/streamfield/blocks/ListBlock.js
@@ -7,7 +7,7 @@ import {
   BaseSequenceChild,
   BaseInsertionControl,
 } from './BaseSequenceBlock';
-import { escapeHtml as h } from '../../../utils/text';
+import { escapeHtml as h } from '../../utils/text';
 
 /* global $ */
