+++ client/src/streamfield/scss/components/c-sf-add-panel.scss
