+++ wagtail/admin/static_src/wagtailadmin/scss/layouts/page-editor.scss
@@ -316,7 +316,7 @@
     }
   }
 
-  // cursory styling for streamfield. Main styling in client/src/components/StreamField/StreamField.scss
+  // cursory styling for streamfield. Main styling in client/src/streamfield/StreamField.scss
   &.block_field {
     padding-inline-start: 20px;
     padding-inline-end: 20px;
