+++ wagtail/admin/static_src/wagtailadmin/scss/panels/streamfield.scss
@@ -1,3 +1,3 @@
 @import '../../../../../../client/scss/settings';
 @import '../../../../../../client/scss/tools';
-@import '../../../../../../client/src/components/StreamField/StreamField';
+@import '../../../../../../client/src/streamfield/StreamField';
